third_party/blink/web_tests/platform/generic/media/media-fragments/TC0051-expected.txt


Title: TC0051
Fragment: 't=3&'
Comment: After processing name-value pairs, this appears to be equivalent to #t=3.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
